Move SQLAlchemy models to /db
