+++ backend/app/db/__init__.py
@@ -1,0 +1,16 @@
+from app.db.artifact import Artifact
+from app.db.link import ExternalLink
+from app.db.machine import Machine
+from app.db.simulation import Simulation
+from app.db.status import Status
+from app.db.variable import SimulationVariable, Variable
+
+__all__ = [
+    "Machine",
+    "Status",
+    "Variable",
+    "SimulationVariable",
+    "Artifact",
+    "ExternalLink",
+    "Simulation",
+]
