+++ backend/app/db/artifact.py
@@ -7,10 +7,10 @@ from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.db.base import Base
-from app.models.mixins import IDMixin, TimestampMixin
+from app.db.mixins import IDMixin, TimestampMixin
 
 if TYPE_CHECKING:
-    from app.models.simulation import Simulation
+    from app.db.simulation import Simulation
 
 
 class Artifact(Base, IDMixin, TimestampMixin):
