+++ backend/app/db/machine.py
@@ -2,8 +2,8 @@ from sqlalchemy import Boolean, String, Text
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.db.base import Base
-from app.models.mixins import IDMixin, TimestampMixin
-from app.models.simulation import Simulation
+from app.db.mixins import IDMixin, TimestampMixin
+from app.db.simulation import Simulation
 
 
 class Machine(Base, IDMixin, TimestampMixin):
