--- backend/app/models/__init__.py
@@ -1,16 +1,0 @@
-from app.models.artifact import Artifact
-from app.models.link import ExternalLink
-from app.models.machine import Machine
-from app.models.simulation import Simulation
-from app.models.status import Status
-from app.models.variable import SimulationVariable, Variable
-
-__all__ = [
-    "Machine",
-    "Status",
-    "Variable",
-    "SimulationVariable",
-    "Artifact",
-    "ExternalLink",
-    "Simulation",
-]
